requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
pillow==10.3.0 (o 11.0.0)
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1